refactor(ir): split IR API modules
